requirements.txt
numpy
pandas
scapy
matplotlib
basemap
python-docx
tqdm
ipinfo